feat(normal): add `balanced-binary-tree` solution
 [balanced-binary-tree](https://leetcode.com/problems/balanced-binary-tree/
