+++ src/solution/normal/mod.rs
@@ -1,3 +1,3 @@
 mod two_sum;
 mod is_palindrome;
-mod maximum_depth_of_binary_tree;
+mod tree;
